Safari-611.1.10/LayoutTests/fast/dom/Window/es52-globals-expected.txt

PASS window.hasOwnProperty("Element") is true
PASS window.hasOwnProperty("x") is true
PASS window.hasOwnProperty("y") is false
PASS window.hasOwnProperty("f") is false
PASS window.__proto__.__proto__.hasOwnProperty("f") is true
PASS window.hasOwnProperty("div") is false
PASS window.__proto__.__proto__.hasOwnProperty("div") is true
PASS window.hasOwnProperty("a") is false
PASS Element is not undefined
PASS x is 1
PASS y is undefined.
PASS f is undefined.
PASS div is undefined.
PASS a is undefined.
PASS successfullyParsed is true

TEST COMPLETE
